Fix some indentation, spacing and sintax issues
