+++ src/modules/Utils.js
@@ -1,7 +1,7 @@
 export const retrieveLocalStorage = (key) => {
   const data = localStorage.getItem(key)
-              ? JSON.parse(localStorage.getItem(key))
-              : [];
+    ? JSON.parse(localStorage.getItem(key))
+    : [];
   return data;
 };
 export const saveLocalStorage = (key, value) => localStorage.setItem(key, value);
